+++ composeApp/src/commonMain/kotlin/org/oltionzefi/splinge/Platform.kt
@@ -1,9 +1,11 @@
 package org.oltionzefi.splinge
 
+import kotlinx.coroutines.CoroutineDispatcher
 import org.oltionzefi.splinge.db.DatabaseDriverFactory
 
 interface Platform {
     val name: String
+    val ioDispatcher: CoroutineDispatcher
     fun shareText(text: String, title: String)
     fun openUrl(url: String)
 }